requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
python-multipart>=0.0.6
imagehash>=4.3.1
Pillow>=9.5.0
pyexiftool>=0.5.5
py-tlsh>=4.7.2
pytesseract>=0.3.10
pdf2image>=1.16.0
pdfminer.six>=20221105
python-docx>=0.8.11
magika>=0.5.0
openai-whisper>=20240930
ffmpeg-python>=0.2.0
torch>=2.0.0
meilisearch>=0.28.0
uuid>=1.30
celery>=5.3.0
redis>=5.0.0
httpx>=0.24.0
chardet>=5.0.0
pyyaml>=6.0.0